C02P20-44 Route events to a Svix application
The events should be sent to a certain application in Svix. This routing
is based on the S3 bucket the SIP was originally uploaded. The mapping
is defined in an env var which is basically JSON as a string. If an S3
bucket is unkown in the mapping, it will mark the  event as 'skipped' and
move on.